# capstoneFood/src/main/java/com/benkhalifa/capstoneFood/model/Food.java
package com.benkhalifa.capstoneFood.model;

import javax.persistence.Column;
import javax.persistence.Entity;
import javax.persistence.GeneratedValue;
import javax.persistence.GenerationType;
import javax.persistence.Id;
import javax.persistence.Table;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@NoArgsConstructor
@AllArgsConstructor
@Entity
@Table(name="foods")
public class Food {
	@Id
	@GeneratedValue(strategy = GenerationType.IDENTITY)
	@Column
	private int id;
	
	@Column
	private String name;
	
	@Column
	private String description;
	
	@Column
	private boolean availability;
	
	@Column
	private int price;
	
	@Column
	private String avatar;
	
	
	
	

}