# capstoneFood/src/main/java/com/benkhalifa/capstoneFood/model/Payment.java
package com.benkhalifa.capstoneFood.model;





import java.sql.Date;

import javax.persistence.Column;
import javax.persistence.Entity;
import javax.persistence.GeneratedValue;
import javax.persistence.GenerationType;
import javax.persistence.Id;
import javax.persistence.Table;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@NoArgsConstructor
@AllArgsConstructor
@Entity
@Table(name="payments")
public class Payment {
	@Id
	@GeneratedValue(strategy = GenerationType.IDENTITY)
	@Column
	private int id;
	
	@Column
	private String  nameOnCard;
	
	@Column
	private Date dateOfPayment;
	
	@Column
	private float totalAmount;
	
	@Column
	private long cardNumber;
	
	
	
	

}